% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_prov.R
\name{commit_prov}
\alias{commit_prov}
\title{Set up a commit for new provenance-tracked files}
\usage{
commit_prov(script_file, tag)
}
\arguments{
\item{git_file}{A valid file name (relative or absolute)}

\item{filetype}{The role of this file within this context: 'input', 'output', 'parent_script', or 'sourced_script'.  Defaults to 'input'.}

\item{nolog}{Should this git provenance information be omitted from the log file? Defaults to FALSE.}
}
\description{
Files flagged as "outputs" from the provenance-tracked script
are given a Git commit ID and a commit message.  This script does not
push the files, only commits them.
}
\examples{
commit_prov()
}

