% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_prov_funs.R
\name{git_prov_funs}
\alias{brick}
\alias{gdal_rasterize}
\alias{git_prov_funs}
\alias{raster}
\alias{rasterize}
\alias{read.csv}
\alias{readOGR}
\alias{readShapePoly}
\alias{read_csv}
\alias{source}
\alias{stack}
\alias{write.csv}
\alias{writeOGR}
\alias{writePolyShape}
\alias{writeRaster}
\alias{write_csv}
\title{Register functions to enable git provenance tracking}
\usage{
source(source_fn, ..., nogit = FALSE)

read.csv(file, stringsAsFactors = FALSE, nogit = FALSE, ...)

write.csv(x, file, row.names = FALSE, nogit = FALSE, ...)

read_csv(file, nogit = FALSE, ...)

write_csv(x, path, nogit = FALSE, ...)

readOGR(dsn, layer, stringsAsFactors = FALSE, nogit = FALSE, ...)

writeOGR(obj, dsn, layer, driver = "ESRI Shapefile", nogit = FALSE, ...)

readShapePoly(fn, nogit = FALSE, ...)

writePolyShape(x, fn, nogit = FALSE, ...)

raster(x, nogit = FALSE, ...)

brick(x, nogit = FALSE, ...)

stack(x, nogit = FALSE, ...)

writeRaster(x, filename, bylayer = FALSE, nogit = FALSE, ...)

gdal_rasterize(..., nogit = FALSE)

rasterize(x, y, filename = "", nogit = FALSE, ...)
}
\description{
Utility functions to assist working with git_prov.  Each of these
reassigns the original function to a new function in the provRmd
namespace.  The new function just adds an automatic call to git_prov()
when appropriate.
}
\details{
To override these functions, use the double-colon notation  (e.g.
utils::write.csv() or readr::read_csv()) to call the function directly
from the original package; alternately, add an argument
of nogit = TRUE to suppress the git_prov() call.
}

