% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prov.R
\name{plot_prov}
\alias{plot_prov}
\title{Access git provenance information}
\usage{
plot_prov(df = .provEnv$script_track, plot_dir = c("TB", "LR")[1])
}
\arguments{
\item{df}{A dataframe of a provenance log file (i.e. \code{.provEnv$script_track}
created by call to \code{script_prov()})}

\item{plot_dir}{Should the graph plot top-to-bottom ('TB') or
left-to-right ('LR')?  Defaults to 'LR'.}
}
\value{
Returns a graph object created by the \code{{DiagrammeR}} package.  This
should be passed to a call to \code{DiagrammeR::render_graph()}
}
\description{
This function takes a filename and reads its git log, strips to its
most recent commit, adds a line to prov_trackfor this file, and returns
a dataframe with git provenance information.
}
\examples{
prov_graph <- plot_prov(.provEnv$script_prov)
DiagrammeR::render_graph(prov_graph)
}

