% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_setup.R
\name{prov_setup}
\alias{prov_setup}
\title{Set up git provenance tracking process}
\usage{
prov_setup(run_tag = "standard run")
}
\arguments{
\item{run_tag}{A character string that briefly describes this particular run. Default = 'standard run'.}
}
\description{
This function sets up a number of global variables used to track provenance
information during the run of the script.  If used within an R Markdown
script, uses the knitted file name and directory.  Sets up NULL variables that
will contain prov_track and script_track info, timing info, and other info
used in the final provenance output dataframe.
}
\examples{
prov_setup()
prov_setup(run_tag = 'testing with new data set')
}

