% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_wrapup.R
\name{prov_wrapup}
\alias{prov_wrapup}
\title{Wrap up provenance tracking}
\usage{
prov_wrapup(include_summary = TRUE, include_workflow = TRUE,
  plot_dir = "TD", include_table = TRUE, commit_outputs = TRUE)
}
\arguments{
\item{include_summary}{Include a summary including session and system info in the rendered Rmd.}

\item{include_workflow}{Create a flow chart of the workflow as determined by
the provenance tracking, and insert into rendered Rmd.}

\item{plot_dir}{Direction of workflow plot: \code{'TD'} (default) creates a top-down plot
while \code{'LR'} creates a left-right plot}

\item{include_table}{Create an output table in the rendered Rmd.}

\item{commit_outputs}{Argument passed to script_prov(); create a commit for any new files created during this run? (default TRUE)}
}
\description{
This function complets the provenance tracking by capturing the provenance
info for the parent script, as well as the session and system info.  This
is saved to the provenance log file.  An optional footer is created (by
default) including a brief session/system summary, workflow chart, and
table of inputs/outputs.  If a footer is desired, this function should be
called inside a code chunk with chunk option: \code{results = 'asis'}
}
\examples{
prov_wrapup()
}

