% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_prov.R
\name{git_prov}
\alias{git_prov}
\title{Access git provenance information}
\usage{
git_prov(git_file, filetype = c("input", "output", "parent_script",
  "sourced_script")[1], nolog = FALSE)
}
\arguments{
\item{git_file}{A valid file name (relative or absolute)}

\item{filetype}{The role of this file within this context: 'input', 'output', 'parent_script', or 'sourced_script'.  Defaults to 'input'.}

\item{nolog}{Should this git provenance information be omitted from the log file? Defaults to FALSE.}
}
\description{
This function takes a filename and reads its git log, strips to its
most recent commit, and returns a dataframe with git provenance information.
If a prov log exists (and `nolog` is FALSE), this information is added to the log.
}
\examples{
git_prov()
}

