% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prov.R
\name{plot_prov}
\alias{plot_prov}
\title{Access git provenance information}
\usage{
plot_prov(df, plot_dir = c("TB", "LR")[1])
}
\arguments{
\item{df}{A dataframe of a provenance log file (from script_prov())}

\item{filetype}{The role of this file within this context: 'input', 'output', 'parent_script', or 'sourced_script'.  Defaults to 'input'.}

\item{nolog}{Should this git provenance information be omitted from the log file? Defaults to FALSE.}
}
\description{
This function takes a filename and reads its git log, strips to its
most recent commit, adds a line to prov_trackfor this file, and returns
a dataframe with git provenance information.
}
\examples{
plot_prov()
}

