% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_wrapup.R
\name{prov_wrapup}
\alias{prov_wrapup}
\title{Wrap up provenance tracking}
\usage{
prov_wrapup(footer = TRUE, workflow = TRUE, table = c("dt", "kable",
  "none")[1], cleanup = TRUE, tag = .prov_run_tag, commit_outputs = TRUE)
}
\arguments{
\item{footer}{Create a footer including session and system info.}

\item{workflow}{Create a flow chart of the workflow as determined by the provenance tracking, and append to the footer (footer must be TRUE)}

\item{table}{Create an output table in the footer; 'dt' creates a DT::datatable() interactive table while 'kable' creates a static table; 'none' creates no table}

\item{cleanup}{Deletes pre-built footers copied in prov_setup().}

\item{tag}{Argument passed to script_prov(); a run tag to be included in provenance log, e.g. 'standard run' (default)}

\item{commit_outputs}{Argument passed to script_prov(); create a commit for any new files created during this run? (default TRUE)}
}
\description{
This function complets the provenance tracking by capturing the provenance
info for the parent script, as well as the session and system info.  This
is saved to the provenance log file.  An optional footer is created (by
default) including a brief session/system summary, workflow chart, and
table of inputs/outputs.
}
\examples{
git_wrapup()
git_wrapup(footer = FALSE, tag = 'debugging data prep', commit_outputs = FALSE)
}

